implemented timer for crewmates win and refactoring web socket connection
