+++ src/app/game/setup/lobby/[gameCode]/page.tsx
@@ -12,9 +12,10 @@ import LobbyReadyToStartText from "./LobbyReadyToStartText";
 import LobbyHeader from "./LobbyHeader";
 import { fetchGame } from "./actions";
 import { Game } from "@/app/types";
+import { useWebSocket } from "@/state/useWebSocket";
 
 export default function Lobby() {
-  const [stompClient, setStompClient] = useState<any>(null);
+  const stompClient = useWebSocket("http://localhost:5010/ws");
   const router = useRouter();
   const { gameCode } = useParams();
   const { game, updateGame } = useGame();
@@ -40,22 +41,8 @@ export default function Lobby() {
   }, []);
 
   useEffect(() => {
-    if (!stompClient) {
-      const socket = new SockJS("http://localhost:5010/ws");
-      const client = Stomp.over(socket);
-      client.connect({}, () => {
-        setStompClient(client);
-      });
-
-      return () => {
-        if (stompClient) {
-          stompClient.disconnect();
-        }
-      };
-    }
-
     loadGameData();
-  }, [stompClient]);
+  }, []);
 
   useEffect(() => {
     if (!stompClient || !gameCode) return;
